+++ content/header.php
@@ -1,10 +1,3 @@
-<?php
-// Load initialization files
-require_once __DIR__.'/init.php';
-
-// Include functions
-require_once PATH.INC.'/functions.php';
-?>
 <!DOCTYPE html>
 <html>
 	<head>