+++ content/index.php
@@ -1,0 +1,9 @@
+<?php
+// Stop execution if the file is accessed directly
+if(!defined('PATH')) exit();
+?>
+<?php getHeader(); ?>
+<div class="wrapper">
+	Hello World!
+</div>
+<?php getFooter(); ?>